+++ app/components/layout/global-footer.tsx
