+++ app/components/sections/hero.tsx
